rename and move some of the converter methods
