+++ slog/converter.go
@@ -1,20 +1,22 @@
-package sloglogrus
+package slog
 
 import (
+	"context"
+	"github.com/samber/lo"
 	"log/slog"
+	"runtime"
+	"slices"
 )
 
-var SourceKey = "source"
-var ErrorKeys = []string{"error", "err"}
+var sourceKey = "source"
+var errorKeys = []string{"error", "err"}
 
-type converter func(addSource bool, replaceAttr func(groups []string, a slog.Attr) slog.Attr, loggerAttr []slog.Attr, groups []string, record *slog.Record) map[string]any
-
-func defaultConverter(addSource bool, replaceAttr func(groups []string, a slog.Attr) slog.Attr, loggerAttr []slog.Attr, groups []string, record *slog.Record) map[string]any {
+func convert(addSource bool, replaceAttr func(groups []string, a slog.Attr) slog.Attr, loggerAttr []slog.Attr, groups []string, record *slog.Record) map[string]any {
 	attrs := appendRecordAttrsToAttrs(loggerAttr, groups, record)
 
-	attrs = replaceError(attrs, ErrorKeys...)
+	attrs = replaceError(attrs, errorKeys...)
 	if addSource {
-		attrs = append(attrs, source(SourceKey, record))
+		attrs = append(attrs, source(sourceKey, record))
 	}
 	attrs = replaceAttrs(replaceAttr, []string{}, attrs...)
 	attrs = removeEmptyAttrs(attrs)
@@ -22,4 +24,193 @@ func defaultConverter(addSource bool, replaceAttr func(groups []string, a slog.A
 	output := attrsToMap(attrs...)
 
 	return output
+}
+
+type replaceAttrFn = func(groups []string, a slog.Attr) slog.Attr
+
+func appendRecordAttrsToAttrs(attrs []slog.Attr, groups []string, record *slog.Record) []slog.Attr {
+	output := slices.Clone(attrs)
+
+	slices.Reverse(groups)
+	record.Attrs(func(attr slog.Attr) bool {
+		for i := range groups {
+			attr = slog.Group(groups[i], attr)
+		}
+		output = append(output, attr)
+		return true
+	})
+
+	return output
+}
+
+func replaceError(attrs []slog.Attr, errorKeys ...string) []slog.Attr {
+	replaceAttr := func(groups []string, a slog.Attr) slog.Attr {
+		if len(groups) > 1 {
+			return a
+		}
+
+		for i := range errorKeys {
+			if a.Key == errorKeys[i] {
+				if err, ok := a.Value.Any().(error); ok {
+					return slog.Any(a.Key, err.Error())
+				}
+			}
+		}
+		return a
+	}
+	return replaceAttrs(replaceAttr, []string{}, attrs...)
+}
+
+func replaceAttrs(fn replaceAttrFn, groups []string, attrs ...slog.Attr) []slog.Attr {
+	for i := range attrs {
+		attr := attrs[i]
+		value := attr.Value.Resolve()
+		if value.Kind() == slog.KindGroup {
+			attrs[i].Value = slog.GroupValue(replaceAttrs(fn, append(groups, attr.Key), value.Group()...)...)
+		} else if fn != nil {
+			attrs[i] = fn(groups, attr)
+		}
+	}
+
+	return attrs
+}
+
+func source(sourceKey string, r *slog.Record) slog.Attr {
+	fs := runtime.CallersFrames([]uintptr{r.PC})
+	f, _ := fs.Next()
+	var args []any
+	if f.Function != "" {
+		args = append(args, slog.String("function", f.Function))
+	}
+	if f.File != "" {
+		args = append(args, slog.String("file", f.File))
+	}
+	if f.Line != 0 {
+		args = append(args, slog.Int("line", f.Line))
+	}
+
+	return slog.Group(sourceKey, args...)
+}
+
+func removeEmptyAttrs(attrs []slog.Attr) []slog.Attr {
+	return lo.FilterMap(attrs, func(attr slog.Attr, _ int) (slog.Attr, bool) {
+		if attr.Key == "" {
+			return attr, false
+		}
+
+		if attr.Value.Kind() == slog.KindGroup {
+			values := removeEmptyAttrs(attr.Value.Group())
+			if len(values) == 0 {
+				return attr, false
+			}
+
+			attr.Value = slog.GroupValue(values...)
+			return attr, true
+		}
+
+		return attr, !attr.Value.Equal(slog.Value{})
+	})
+}
+
+func attrsToMap(attrs ...slog.Attr) map[string]any {
+	output := map[string]any{}
+
+	attrsByKey := groupValuesByKey(attrs)
+	for k, values := range attrsByKey {
+		v := mergeAttrValues(values...)
+		if v.Kind() == slog.KindGroup {
+			output[k] = attrsToMap(v.Group()...)
+		} else {
+			output[k] = v.Any()
+		}
+	}
+
+	return output
+}
+
+func groupValuesByKey(attrs []slog.Attr) map[string][]slog.Value {
+	result := map[string][]slog.Value{}
+
+	for _, item := range attrs {
+		key := item.Key
+		result[key] = append(result[key], item.Value)
+	}
+
+	return result
+}
+
+func mergeAttrValues(values ...slog.Value) slog.Value {
+	v := values[0]
+
+	for i := 1; i < len(values); i++ {
+		if v.Kind() != slog.KindGroup || values[i].Kind() != slog.KindGroup {
+			v = values[i]
+			continue
+		}
+
+		v = slog.GroupValue(append(v.Group(), values[i].Group()...)...)
+	}
+
+	return v
+}
+
+func appendAttrsToGroup(groups []string, actualAttrs []slog.Attr, newAttrs ...slog.Attr) []slog.Attr {
+	actualAttrs = slices.Clone(actualAttrs)
+
+	if len(groups) == 0 {
+		return uniqAttrs(append(actualAttrs, newAttrs...))
+	}
+
+	for i := range actualAttrs {
+		attr := actualAttrs[i]
+		if attr.Key == groups[0] && attr.Value.Kind() == slog.KindGroup {
+			actualAttrs[i] = slog.Group(groups[0], lo.ToAnySlice(appendAttrsToGroup(groups[1:], attr.Value.Group(), newAttrs...))...)
+			return actualAttrs
+		}
+	}
+
+	return uniqAttrs(
+		append(
+			actualAttrs,
+			slog.Group(
+				groups[0],
+				lo.ToAnySlice(appendAttrsToGroup(groups[1:], []slog.Attr{}, newAttrs...))...,
+			),
+		),
+	)
+}
+
+func uniqAttrs(attrs []slog.Attr) []slog.Attr {
+	return uniqByLast(attrs, func(item slog.Attr) string {
+		return item.Key
+	})
+}
+
+func uniqByLast[T any, U comparable](collection []T, iteratee func(item T) U) []T {
+	result := make([]T, 0, len(collection))
+	seen := make(map[U]int, len(collection))
+	seenIndex := 0
+
+	for _, item := range collection {
+		key := iteratee(item)
+
+		if index, ok := seen[key]; ok {
+			result[index] = item
+			continue
+		}
+
+		seen[key] = seenIndex
+		seenIndex++
+		result = append(result, item)
+	}
+
+	return result
+}
+
+func contextExtractor(ctx context.Context, fns []func(ctx context.Context) []slog.Attr) []slog.Attr {
+	var attrs []slog.Attr
+	for _, fn := range fns {
+		attrs = append(attrs, fn(ctx)...)
+	}
+	return attrs
 }
